requirements.txt
# Core dependencies
discord.py[voice]>=2.3.2
python-dotenv>=1.0.0
PyYAML>=6.0.1

# Database
aiosqlite>=0.19.0

# Date/Time
python-dateutil>=2.8.2

# Text-to-Speech
edge-tts>=6.1.9
PyNaCl>=1.5.0  # Required for voice support

# Development tools
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
types-PyYAML>=6.0.12.12
types-python-dateutil>=2.8.19.14

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1

# Linting
isort>=5.12.0